优惠券数量统计和status字段类型修改

优惠券数量统计和status字段类型修改

See merge request !338

+++ marketing-tool-po/src/main/java/cn/rongcapital/mkt/po/MaterialCoupon.java
@@ -34,7 +34,7 @@ public class MaterialCoupon extends BaseQuery{
 
     private String taskName;
 
-    private Integer status;
+    private Byte status;
 
     private Date createTime;
 
@@ -151,12 +151,12 @@ public class MaterialCoupon extends BaseQuery{
     public void setTaskName(String taskName) {
         this.taskName = taskName == null ? null : taskName.trim();
     }
-
-    public Integer getStatus() {
+    
+    public Byte getStatus() {
         return status;
     }
 
-    public void setStatus(Integer status) {
+    public void setStatus(Byte status) {
         this.status = status;
     }
